requirements.txt
redis
streamlit
python-dotenv
google-generativeai
pymupdf
pathlib